requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.36
pyodbc==5.2.0
python-dotenv==1.0.1
